chore: Update the `tx_busy_handler()` in stacks-common also
